backend/requirements.txt
Django
djangorestframework
celery
redis
drf_yasg
djangorestframework-simplejwt
django-rest-swagger
django-cors-headers
